First full working version 2

+++ classes/UniqueMediaFinder.php
@@ -2,8 +2,11 @@
 namespace SNiPI\UniqueMediaFinder\Classes;
 
 use Illuminate\Filesystem\Filesystem;
+use Storage;
 use Backend\Widgets\MediaManager;
 use Input;
+use Mail;
+use Lang;
 use BackendAuth;
 use SNiPI\UniqueMediaFinder\Models\Search;
 use SNiPI\UniqueMediaFinder\Models\MetadataInformations;
@@ -54,6 +57,17 @@ class UniqueMediaFinder {
                 
 			});
 
+			$widget->addDynamicMethod('onFeedback', function() use ($widget){
+
+				Mail::raw([
+				    'html' => '<p><strong>'.Input::get('your_name') . '</strong> sent feedback form from plugin.</p><p><strong>Feedback message:</strong><br/>' . Input::get('your_message'). '</p><p><strong>Contact:</strong><br/>' . Input::get('your_email').'</p>'
+				], function ($message) {
+					$message->to('snipiba@gmail.com','Snipi');
+					$message->subject('Feedback from Unique Media Finder plugin');
+				    //
+				});
+			});
+
 			$widget->addDynamicMethod('onLoadFilters', function() use ($widget){
 
 				$widget->vars['provider'] = Input::get('provider');
@@ -73,10 +87,12 @@ class UniqueMediaFinder {
 					
 				}
 
-				$data = $this->providers[$provider]->search($qs, $options);
+				$data = $this->providers[$provider]->search($qs, $options, Input::get('page') ?? 1);
 				$widget->vars['items'] = $data['results'];
 				$widget->vars['data'] = $data;
 				$widget->vars['provider'] = $provider;
+				$widget->vars['search'] = $qs;
+				$widget->vars['page'] = Input::get('page') ?? 0;
 				$widget->vars['currentProvider'] = $this->providers[$provider];        
 
 				if(Settings::get('store_search')) {
@@ -91,6 +107,14 @@ class UniqueMediaFinder {
 
 			});
 
+			$widget->addDynamicMethod('onPaginate',function() use($widget){
+				$qs = Input::get('query');
+				$page = Input::get('page');
+				$provider = Input::get('provider');
+
+
+			});
+
 			$widget->addDynamicMethod('onShowDetails', function() use ($widget){
 				$provider = Input::get('provider');
 				$id = Input::get('id');
@@ -99,6 +123,10 @@ class UniqueMediaFinder {
 				$widget->vars['provider'] = $provider;
 				$widget->vars['id'] = $id;
 				return $widget->makePartial($provider .'/details', ['photo' => $data]);
+			});
+
+			$widget->addDynamicMethod('onAbout', function() use ($widget){
+				return $widget->makePartial('editor/' . Lang::getLocale().'/about');
 			});
 
 			$widget->addDynamicMethod('onDownload', function() use ($widget){
@@ -119,15 +147,48 @@ class UniqueMediaFinder {
 
 				$widget->bindEvent('file.rename', function ($originalPath, $newPath) {
 			        // Update custom references to path here
+
 			        $origFile = MetadataInformations::where('filename', basename($originalPath))->first();
-			        $origFile->filename = basename($newPath);
-			        $origFile->save();
+			        if($origFile) {
+				        $origFile->filename = basename($newPath);
+				        $origFile->full_path = $newPath;
+				        $origFile->save();
+				    }
+			    });
+
+				$widget->bindEvent('file.move', function ($originalPath, $newPath) {
+					$filename = basename($originalPath);
+			        $origFile = MetadataInformations::where('filename', basename($originalPath))->first();
+			        if($origFile) {
+				        $origFile->filename = basename($originalPath);
+				        $origFile->full_path = $newPath.'/'.$filename;
+				        $origFile->save();
+				    }
 			    });
 
 				$widget->bindEvent('file.delete', function ($originalPath) {
 			        // Update custom references to path here
 			        $origFile = MetadataInformations::where('filename', basename($originalPath))->first();
-			        $origFile->delete();
+			        if($origFile) {
+				        $origFile->delete();
+				    }
+			    });
+
+
+				$widget->bindEvent('folder.rename', function ($originalPath, $newPath) {
+
+			        $items = MetadataInformations::where('full_path','like', '%'.$originalPath.'%')->get();
+			        foreach($items as $item) {
+			        	$item->full_path = str_replace($originalPath, $newPath, $item->full_path);
+			        	$item->save();
+			        }
+			        
+			    });
+
+				$widget->bindEvent('folder.delete', function ($originalPath) {
+			        // Update custom references to path here
+			        $items = MetadataInformations::where('full_path','like', '%'.$originalPath.'%')->delete();
+			        
 			    });
 
 			}
